world: probe index counter uint16 -> uint32
_probe_idx counts probes drawn per (agent, family) and seeds the next instance index. uint16 caps it at 65,535:
sequential_halving at n = 10^7, b = 3 hands the final winner the budget remainder (146,452 pulls) and overflowed,
so all 500 units at that rung lack the halving arm. MIDIAN-SH's '-10 out of bounds for uint16' was the same counter.
No existing result changes -- the instance sequence is identical, there is just no cap. 640 MB instead of 320 MB at
n = 10^7.

+++ rte/world.py
@@ -256,7 +256,7 @@ class World:
         self.D = apply_lying(D_honest, self.liars, lie_mode, self.demand)
         self.D_view = self.D.copy(); self.D_view.setflags(write=False)
         self._obs: dict[int, dict[int, list]] = {}    # reporter j -> {a: [sum, cnt]} (scalar report path)
-        self._probe_idx = np.zeros((self.n, self.K), np.uint16)    # probes drawn so far per (agent, family)
+        self._probe_idx = np.zeros((self.n, self.K), np.uint32)    # probes drawn so far per (agent, family); uint16 capped at 65,535 and halving hands one winner 146k at n=1e7 b=3
         self._probe_salt = stable_seed_32(seed, "probes")
         self.epoch = np.zeros(self.n, np.int32); self.seen_epoch = np.zeros(self.n, np.int32); self.churn_events = 0
         self._snap = (self.S.copy(), self.D.copy(), self.liars.copy(), self.backend.snapshot())
